Add command line args
Because:
- The experiment was not configurable from the command line

This commit allows to define the configuration in the NamedTuple objects
and automatically adds all fields from it to an argparser. They can be then overwritten
from the command line by simply calling --arg value. Care must be taken with boolean flags,
which currently use the eval() function as type conversion. This MIGHT lead to unexpected
behaviour or produce non-clear error messages.